Inital implementation of Burges second spell.

+++ Assets/Scripts/ScriptableObjects/SpellData/Burge/BurgeSpell1Data.cs
@@ -3,7 +3,9 @@ using System.Collections.Generic;
 using UnityEngine;
 
 /*
-* Purpose: Extends SpellData for spell specific constant data.
+* Purpose: Implements Bahri'a first spell. Bahri throws a damage dealing orb at a target direction that returns after reaching its maximum magnitude.
+* Any unit hit on the initial cast takes magic damage and any unit hit by the return takes true damage.
+* The orb accelerates upon starting its return until reaching Bahri.
 *
 * @author: Colin Keys
 */
@@ -14,5 +16,5 @@ public class BurgeSpell1Data : SpellData
     [field: SerializeField] public float jumpTime { get; private set; }
     [field: SerializeField] public float hitboxLength { get; private set; }
     [field: SerializeField] public float hitboxWidth { get; private set; }
-    [field: SerializeField] public float numberOfHits { get; private set; }
+    [field: SerializeField] public int numberOfHits { get; private set; }
 }
